requirements.txt
asgiref==3.11.0
attrs==25.4.0
dj-rest-auth==7.0.2
Django==5.2.10
django-allauth==65.14.0
django-cors-headers==4.9.0
django-extensions==4.1
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
drf-yasg==1.21.14
importlib_resources==6.5.2
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
packaging==26.0
pillow==12.1.0
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
rest-framework-simplejwt==0.0.2
rpds-py==0.30.0
sqlparse==0.5.5
swagger-spec-validator==3.0.4
typing_extensions==4.15.0
uritemplate==4.2.0
